Add required paragraph field to library.properties
Missing paragraph field causes the Arduino IDE to display a warning whenever the library is compiled or Library/Boards Manager is opened/closed:

Invalid library found in E:\arduino\libraries\serde: Missing 'paragraph' from library

This also results in the library's examples not being shown in the File > Examples menu. With older versions of the Arduino IDE, it caused compilation of the library to fail.